chore(repo): add tsconfig path handling to generate CLI docs
