+++ packages/tao/src/shared/package-json.ts
@@ -2,19 +2,19 @@ import { TargetConfiguration } from './workspace';
 
 export type PackageJsonTargetConfiguration = Omit<
   TargetConfiguration,
-  'executor' | 'dependsOn'
+  'executor'
 >;
 
 export interface NxProjectPackageJsonConfiguration {
-  targets: Record<keyof PackageJson['scripts'], PackageJsonTargetConfiguration>;
+  targets: Record<string, PackageJsonTargetConfiguration>;
 }
 
 export interface PackageJson {
   name: string;
-  scripts: Record<string, string>;
-  dependencies: Record<string, string>;
-  devDependencies: Record<string, string>;
-  peerDependencies: Record<string, string>;
+  scripts?: Record<string, string>;
+  dependencies?: Record<string, string>;
+  devDependencies?: Record<string, string>;
+  peerDependencies?: Record<string, string>;
   nx: NxProjectPackageJsonConfiguration;
 }
 
@@ -27,8 +27,8 @@ export function buildTargetFromScript(
   return {
     executor: '@nrwl/workspace:run-script',
     options: {
-      script,
       ...(nxTargetConfiguration.options || {}),
+      script,
     },
     ...nxTargetConfiguration,
   };
